refactor: AnimationCircle 컴포넌트 분리
